fix check for "simple" loop termination condition
If the termination condition of a for loop only involves upper bounds,
then we use a simplified method of constructing the corresponding iteration
domain.  Unfortunately, our check for this property was faulty.
In particular, it would mark the termination condition of the following
loop as only involving upper bounds.

		for (j = i; j <= 5 * ((j + 2)/5); ++j)

The simplified method cannot be used on this loop because
the termination condition has an implicit lower bound through
the integer division.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>