avoid wrapping on some simple cases of loops with unsigned iterators
Also applying wrapping even when it is not necessary always produced
correct the results, the representation may be overly complicated.

We need to update some of the test cases because in the case of
wrapping, the schedule needs to include the constraints of the domain,
while they are not included when wrapping is not applied.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>