accept non-affine conditions in if statements
If the condition is non-affine, an extra statement is constructed
that evaluates the conditions and writes the result to a virtual
scalar.  This scalar is then added to the list of arguments of
each statement depending on the condition and the domains of the
statements are adjusted to have a value of 1 (in the then branch)
or 0 (in the else branch) for this argument.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>