turn virtual scalars into virtual arrays
There is no point in requiring a dependence analysis to find
out which statement instance wrote the value of the virtual scalar
for a given iteration of a depending statement.
During construction, we know which statement instance is involved,
so we may just as well use that instance as the index into the
virtual array.  Note that this makes the virtual arrays single
assignment.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>