allow nested accesses in if conditions
Before, we would treat any non-static condition as a non-affine
condition, introduce an extra statement to evaluate the condition
and represent the result by a boolean "argument" of the iteration domain.
This hides the condition inside that statement and therefore makes
is difficult to figure out any possible relations between different
conditions in the code.

We now also allow iteration domain arguments to refer directly to any
nested access in an otherwise affine condition.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>