induction variable of for loop should be considered a parameter inside the loop
Before, we would mark the induction variable as having an unknown
value, even during the analysis of the loop and even if no assignments
were being performed to that variable inside the loop body.

Instead, we should consider the variable as a parameter during the
analysis of the loop (unless it is assigned somewhere inside the
loop body).  Only after the loop should we consider the induction
variable to have some unknown value.  In principle, we can turn
this into a known value in some cases.  We may consider that later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>