extract_affine: a variable that has been assigned an unknown value is not affine
An affine expression can only involve iterators and parameters.
During the analysis of a loop body, iterators are treated as parameters,
so we only need to consider parameters.
If a given variable has been assigned a value inside the scop,
then it cannot be a parameter.
If the value assigned is known, then we may still be able to recover
some affine expression, but if it is not, then we should consider
the expression to be non-affine.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>