an address passed to a const pointer argument should not be considered a write
If the address of some scalar is taken, we assume it may be written.
However, if this address is simply passed to a function through an
argument of const pointer type, then there is no need to do so.
Especially since elsewhere in the code, we consider such constructs
to be a read from the scalar rather than a write.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>