destroy MyDiagnosticPrinter on newer clangs
In 31e1c8e (adapt to the recent changes in clang,
Thu Oct 6 12:43:12 2011 +0100), a clone method was added to
MyDiagnosticPrinter to support recent clangs, but we didn't take
into account that if the MyDiagnosticPrinter, then the original
MyDiagnosticPrinter was left hanging around.  We now destroy it
if it turns out that the clone method has ever been called.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>