cache values in assigned_value instead of expressions
The value of an expression may change between the time it was
assigned to a variable and the time this variable is used.
Evaluating the expression at the time the variable is used may
therefore produce incorrect results.

Furthermore, if the value refers to any parameters, then we
should discard the value if one of these parameters is assigned
a value (and therefore turns out not to be a parameter).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>