support break and continue
To be able to handle breaks and continues, we keep track of the condition
under which subsequent statements in the current loop iteration
or any statement in later loop iterations should not be executed.
These "skip" conditions may be either affine constraints or they
may be represented by the value of a virtual array.
The conditions, when introduced by a break or continue statement,
start out as universal constraints and get modified when the statement
is embedded in an if statement and/or combined with other statements.
These modified skip conditions may require the introduction of additional
virtual arrays to represent the combination of the skip conditions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>