PetScan::extract_for: explicitly keep track of when iterator is virtual
The wrapping map is only computed when we use a virtual iterator,
so we need to make sure that we are in fact using a virtual iterator
before we use that map.
The original condition was probably correct, but not very obvious.
Make it explicit so that we can more easily change the code later.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>