pet_scop_extract_from_C_source: reset user pointers on parameter ids in result
For each parameter, pet stores a reference to the corresponding ValueDecl *
in the user pointer of its isl_id.  When the ASTContext is destroyed, these
pointers no longer have a valid value.  Moreover, in interactive applications,
we want the user to be able to refer to the parameter simply by name.
Since isl_ids with different user pointers are treated as being distinct,
this requires the isl_ids to have a NULL user pointer.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>