add test for verifying the correctness of the isl code generator on its tests
Ths isl code contains test cases for testing the correctness of the isl
code generator.  However, the output of the code generator is simply
compared to a previously computed output.  The correctness of those
previously computed outputs is verified by this commit.

In particular, for each of the test cases, we first generate code
use the isl code generator in pet_codegen.  We then parse
the output using pet in pet_check_code and check if the code
satisfies the input schedule.

In particular, we check that
- the domains are identical, i.e., the calls in the C code
  correspond to the domain elements of the schedule
- the calls are performed in an order that is compatible
  with the schedule
- no function is called twice with the same arguments, provided
  the schedule is single-valued

We only perform these tests if we can find the isl source code tree.
That is, we do not perform the tests if we are using a "system" isl.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>