scop_add_break: try and merge filters
For statements that follow a break inside a loop, the skip condition
corresponding to the break may be applied twice, once because it follows
the break inside the same iteration and once because it may be executed
in a later iteration.  In such cases, we want to combine the two filters
into a single filter that applies to the current and to previous iterations.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>