handle generic conditions is for loops
They are handled in essentially the same way as while loops,
except that we need to take into account the loop iterator on the for loop.

Because we don't know in advance if we are going to need a separate
statement to evaluate the loop condition, we may need to make room
for such a statement even when we turn out not to need it.
This commit may therefore also affect the statement numbers
in some cases.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>