pet_scop_embed: take mapping from virtual to real iterator
In the next commit, we will add support for generic conditions in
for loops and then we may not always be able to rewrite an iteration
domain that was constructed in terms of a virtual iterator to one
that refer to th real iterator.  We therefore need to be able to
adjust any reference to the real iterator to one that refers to
the virtual iterator.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>