PetScan::extract_for: also check whether bound is simple before computing gist
The gist operation was introduced in 2fcd426 (PetScan::extract_for: compute
gist before testing whether bound is simple, Thu Jan 19 14:46:07 2012 +0100)
so that we could ignore bounds that were implied by the initial value.
However, the gist operation itself may introduce equalities if they
are implied by inequalities in the original set and the context.
Also check for a simple bound before computing the gist.  This also
saves us from computing the gist if the condition is already simple.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>