pet_scop: keep track of start and end file offsets of the extracted scop
This should make life easier on users of pet that want to replace
the original scop by transformed code.  In particular, if the scop
was detected using --autodetect then it was nearly impossible
for a user to figure out which part of the input corresponds to the scop.

We do not verify the values of these offsets in our test cases
because the exact offsets may be slightly different for different
versions of clang.  In particular, we use some utility functions
in newer versions of clang and try to mimic them for older versions
of clang.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>