derive offset information from location of scop and endscop pragmas
The file offsets derived from the AST currently do not take into account
leading spaces on the first statement and (in some cases) trailing
spaces and newline on the last statement.
In the presence of scop and endscop pragmas, we can simply make the offsets
refer to the piece of text delimited by the two pragmas.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>