pet_check_code: better support for schedules that are not single valued
The check for whether the schedule order is respected by the generated code
could fail on correct code if the schedule is not single valued because
different schedulings of the same statement instance may be in different
positions compared to other statement instances.

We include a different test that should work in general, but we only
apply it when the schedule is not single valued since it is considerably
more expensive.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>