stop destroying MyDiagnosticPrinter
c8d4ee7 (destroy MyDiagnosticPrinter on newer clangs,
Sat Nov 12 16:46:12 2011 +0100) added code that tried to delete
the MyDiagnosticPrinter object on the versions of clang where
createDiagnostics would clone the object instead of taking ownership.
In d40dc1f (fix use of DiagnosticOptions for older clangs,
Thu Nov 8 13:10:09 2012 +0100) we replaced the call to createDiagnostics
by a call to setClient on the DiagnosticsEngine created by
the CompilerInstance.  Since setClient takes ownership of the client
by default, the delete should no longer be necessary and in fact
leads to double frees on some clang versions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>