keep track of index expressions
Until now, we only kept track of an "access relation" that was used
(by outside users) both for dependence analysis and for printing statement
bodies.  For the latter purpose, an explicit function representation
through an "index expression" is more convenient as it can more closely
match the original code.  In particular, when passing arrays or array
slices to a function, an index expression can keep track of how many
indices were specified in the original code.

At this stage, index expressions and access relations usually contain
the same information so we may be duplicating effort with little gain.
In future, however, access relations and index expressions may diverge
further, in particular when we start keeping track of which elements
of an array are accessed by a function call.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>