use an isl_multi_pw_aff to represent a skip condition
In particular, use either an affine expression or an access to a variable.
Note that this commit not only changes the representation from an isl_set
to an isl_multi_pw_aff, it also changes the space in which these objects
live.  The new representation has a zero-dimensional domain rather than
a parametric domain since that is what most users expect in the end.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>