PetScan::scan: fix check for scop being proper part of statement
In particular, then end of the scop points beyond the scop,
while the end of the statement that it is compared with points
to the start of the token that represents the end of the statement.
This has only become a problem since e00885c (derive offset information
from location of scop and endscop pragmas, Tue Feb 5 17:37:13 2013 +0100).
Before that commit, the end of the scop would point to the line
containing the endscop pragma, while after that commit, it would
point to the next line.

Reported-by: Athanasios Konstantinidis <ak807@doc.ic.ac.uk>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>