scan.cc: scop_add_while: take isl_id instead of isl_map
All we really need from the isl_map is the output tuple id,
so it makes more sense to pass that id than to take an isl_map
and impose some fairly arbitrary additional constraints on that map.
This should also make it easier to convert some of the maps to functions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>