introduce pet_implication
We want to keep track of the index expressions of all accesses
in a statement body as an explicit function to ease printing
of statement bodies.
Some filters may however access more than one element to indicate
that all the accessed filter elements should have a value
satisfying the filter value relation.
Such filter access relations cannot be represented by an explicit function.
In the next commit, we will therefore break up these filter access relations
into a filter access function and an associated "implications" mapping
to cover the remaining elements in the original filter access relation.
In this commit we introduce the infrastructure for representing these
implications.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>