use implications to encode while and break filters
Filters that express while or break conditions used to be encoded
by filter maps mapping to all previous iterations.  Since we want
to keep track of index expressions in the future, this commits
breaks such filter maps into two parts, a filter expressing the condition
on the previous iteration and an implication expressing that this condition
extends to all previous iterations.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>