add pet_expr_from_index_and_depth wrapper
If an index expression is extracted from the source code, then the
number of indices may be smaller than the depth of the array.
In such cases, the pet_expr_from_index will not construct the correct
access relation.  We therefore add an additional wrapper that takes
an index expression and the depth of the array.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>