pet_scop_collect_{reads,writes}: take arguments into account
The original functions did not take into account that statements and
accesses may have arguments that make the statement or access depend
on values only known at run-time.  That is, we would assume that all
accesses are static and would silently drop accesses in case of such
arguments.

Instead, we now allow for the collection of "may" writes, where the
dynamic values are projected out and of "must" writes, where dynamic
accesses are ignored.  Since there is usually no need to know about
any "must" reads, we only provide a function to collect the "may" reads.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>