pet_scop: keep track of schedule tree
This schedule tree is meant to replace the schedule pieces scattered
over the statements.  The latter will be removed in a subsequent commit.
The union of these pieces forms a flat schedule that can be recovered
from the schedule tree using isl_schedule_get_map.

We produce a schedule tree such that the original schedule extracted
by pet can be used in the same way as a schedule constructed
by isl_schedule_constraints_compute_schedule.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>