pet_check_code: op_expr_extract_pw_aff: handle more types of expressions
In the next commit we will stop trying to convert conditional operators
to affine expressions during the first phase of parsing.
This means that if such an operator is used outside an index expression,
then it will no longer be converted to an affine expression.
Since the isl AST generator may print variable assignments that
include conditional operators, we need to handle any type of expressions
that can appear as arguments to such an operator.

Note that further down the refactoring, we will be able to reuse
pet_expr_extract_affine from pet_check_code.c.  We therefore only
add support for those operations that are needed to pass the tests.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>