PetScan::extract_expr(ConditionalOperator *): preserve original condition
In particular, do not try to convert the condition to an affine expression.
The condition will be converted to affine expression when needed,
i.e., when the entire conditional expression is converted to
an affine expression.

As a result, conditional operators that appear outside of
an index expression are no longer converted to an affine expression.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>