make pet_expr objects reference counted
This will make it easier to keep track of nested accesses as pet_expr
objects.

In order to avoid outside users subverting the reference counting,
we hide the internal representation of pet_expr objects.
We also include a reference to the isl_ctx in whith the pet_expr
is created.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>