PetScan::scan(FunctionDecl *): construct a pet_context up front
Rather than keeping track of assigned values in assigned_values and
converting this map into a pet_context whenever we need one,
we create a pet_context up front and keep track of the assignments
in the pet_context.  This will allow us in subsequent commits
to first extract a pet_tree and to then convert this pet_tree
to a pet_scop outside of the PetScan object.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>