PetScan::resolve_nested: handle self dependences in conditional assignments
In the next commit we will extract access expressions as pet_expr
objects first.  This has the side effect that nested accesses
are directly extracted as arguments of the nesting access
rather than first being represented by parameters.
The conditions of the conditional assignment therefore also
get applied to these nested accesses, whereas they currently are not
since these nested accesses are hidden inside parameters.
This is arguably the correct thing to do, but it may result
in nested accesses depending on themselves if they happen
to also appear in the condition.  We therefore add some special handling
to detect such cases.  This handling will only be effective after the next
commit.  We apply it first in order to make the next commit more
readable.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>