PetScan::extract_access_expr: delay plugging in the affine expression arguments
That is, plug them in where we require them to have been plugged in.
In particular, if we want to extract an affine condition from a pet_expr,
then we had better plug in affine arguments first such that they are
not recognized as nested accesses.
We also need to plug in affine arguments before resolving nested
expression in order to be able to recognize subaccesses.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>