PetScan::extract_non_affine_condition: remove redundant resolve_nested
Commit e8b1fa6 (PetScan::extract_non_affine_condition: resolve nested accesses,
Fri May 18 13:03:19 2012 +0200) introduced two calls to resolve_nested,
one on the expression and one on the scop.  Since resolve_nested
on a scop handles nested accesses in the iteration domains and
the iteration domain is a universe set at this point, the second call
appears to have been introduced by mistake.  Remove it.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>