tree2scop.c: is_nested_allowed: take tree instead of scop
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up to a top-down process.  This means that we cannot
use the pet_scop objects corresponding to child nodes to determine
what to do with the current node.  Use the original pet_tree instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>