tree2scop.c: scop_from_affine_for: delay intersection with affine break
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up to a top-down process.
In particular, we want to pass down the domain information derived
from outer constructs to inner constructs.  This domain information
cannot contain any information based on affine breaks since those
breaks are part of the inner constructs.  We will therefore have to
first construct the domain without affine breaks, then construct
a scop for the body and finally adjust the domain based on the affine
breaks.
Here we perform the first step towards this reorganization.
In particular, we split off the intersection with the affine breaks.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>