tree2scop.c: scop_from_non_affine_while: delay construction of body pet_scop
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up to a top-down process.
In particular, we want to pass down the domain information derived
from outer constructs to inner constructs.  This means we have
to construct the pet_scop for the body after the domain has been constructed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>