pet_stmt_from_pet_expr: allow specification of statement domain
We want to change the extraction of a pet_scop from a pet_tree
from a bottom-up process to a top-down process.
Whereas, currently, the domain of a statement starts out as
an anonymous 0D universe set that is modified as we move up
the pet_tree, in the top-down process, the domain will be known
by the time this function is called.

We allow the users to pass this domain to pet_stmt_from_pet_expr.
In the top-down process, the constraints derived from the outer
constructs will be stored in the pet_context.
We therefore obtain the domain from the pet_context whenever
it is already available and take the current 0D universe set
as input otherwise.  The latter will be changed to take
the domain from the pet_context in subsequent commits.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>