PetScan::extract_scop: clear writes at outer level
Before this commit, we would only clear writes inside the bodies
of other constructs.  This means that variables that get written
somewhere inside the tree may initially get treated as parameters.
This is not so much of a problem now, but in a subsequent commit,
we will start plugging in affine expressions and then we do not
want these variables to be treated as parameters as they would
then incorrectly be considered affine.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>