pet_expr_extract_affine: exploit domain information while extracting divs/mods
An integer division in C is represented by a piecewise affine expression
because the result is different for negative and positive values of the lhs.
If we know that this lhs is positive (or negative), then we can avoid
adding extra pieces, so it is worthwhile to try and simplify the extracted
affine expression in the context of the outer iterators domain.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>