introduce pet_loc
There are currently two places where location information is stored.
We store the line number in a pet_stmt and the start and end offsets
in a pet_scop.  This information is extracted directly from the clang
AST.  In future, we want to extract a pet_scop from an intermediate
pet_tree so we need to attach both pieces of information to a pet_tree.
Combine these two pieces of information into a pet_loc.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>