add encapsulate_dynamic_control option
With this option turned on, any dynamic control is encapsulated
in a macro statement.

We currently detect the dynamic control on our way back up the pet_tree.
This means that we first create scops that may contain dynamic control and
then discard and replace them with a single statement scop.
This process could be optimized by detecting some dynamic control
beforehand.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>