handle pencil independent pragmas
For each pencil independent pragma, we keep track of the pairs
of statement instances that are excluded from depending on each other
as well as the variables that are local to the independent loop.
The representation may not be very efficient, but this potential
problem can be solved when we move to schedule trees and/or
perform dependence analysis inside pet.

We conservatively only keep track of independences on affine for loops
with an affine loop condition.  We can later extend this to other
types of loops as the need arises.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>