separate out extraction of nested expression from isl_id
Nested accesses are currently represented by a clang::Expr object.
We will change this to pet_expr objects and during a transitional phase
a nested access may be represented by either a clang::Expr object
or a pet_expr object.

We separate out the extraction of a nested expression so that we
only have to modify it in one place.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>