pet_expr_is_scalar_access: do not consider scalar fields as scalars
pet_expr_is_scalar_access is used to check whether a value assigned
to a variable should be stored in assigned_value.  If this variable
is the field of a struct, then the value will actually end up being
assigned to the entire struct.  This does not cause any problems
in the current implementation because the value of a struct is
never retrieved from assigned_value.
In future commits, we will be changing the way we keep track
of assigned values, so we need to make sure this will not
cause any problems then.  We therefore no longer consider scalar
members as scalars.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>