drop pet_context_mark_assigned
In the previous commit we started clearing writes at the outermost level.
This means we will no longer come across a variable that is first
treated as a parameter which then later turns out not to be a parameter.
We can therefore remove the special handling of this situation and
simply replace the calls to pet_context_mark_assigned by
calls to pet_context_mark_unknown.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>