PetScan::extract_scop: populate pet_context with parameter assignments
We currently create a new parameter inside pet_expr_extract_affine
whenever we come across a read on an integer variable that has not
been written to anywhere in the tree.  To be able to distinguish
variables that have been not been assigned any value and those
that have been assigned a (known or unknown) value, we have to
keep track of not only the known affine values, but also "unknown" values.

Now that we construct a pet_scop from the top down, we know in
advance which variables should be considered as parameters.
We add them directly to the pet_context before extracting the pet_scop
and its arrays.  In subsequent commits, we can then remove
the creation of parameters on-the-fly and the special treatment
of "unknown" values.  This in turn will then allow us to
reuse pet_expr_extract_affine from pet_check_code.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>