pet_expr_is_equal: compare index expressions up to reordering of parameters
isl_multi_pw_aff_plain_is_equal currently only considers its two
arguments to be obviously equal if they have their parameters
in the same order.  We do not want to depend on this order while
comparing pet_expr objects, especially since we may be changing
the order of the parameters in the following commits.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>