pet_context: no longer keep track of "unknown" values
The only reason we kept track of "unknown" values in the pet_context
was to be able to distinguish variable that have had some value
assigned to them and those that have never been assigned a value.
This information was used inside pet_expr_extract_affine to decide
whether we should consider the variable as a parameter.
Now that this logic has been removed from pet_expr_extract_affine,
there is no longer any need to keep track of unknown values.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>