export pet_expr_extract_affine
Now that we have given the callers of pet_expr_extract_affine a better
control over what should be considered a parameter, we can expose
this function for external use.
Note that we currently do not expose any functionality for setting or
updating any assignemnts in a pet_context.  We can do so later
if the need arises.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>