allow specification of summary functions
A summary function describes the accesses performed by
a function called from a scop and can be used when the original
function body is not available or if it is too complicated
to be analyzed in an accurate way.

Get the summary function supplied by the ``attribute'' pencil_access.
Handling a new attribute in Clang seems to require modifications in
Clang itself.  To avoid this, we define pencil_access as a builtin macro
that expands to the already existing annotate attribute.

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>