PetScan::resolve_nested: handle presence of arguments in access expressions
The original implementation would assume that access expressions
do not have any arguments and would simply populate the list of
arguments from nested accesses.
When we change the extraction of access expressions to not convert
the indices to affine expressions first, the access expressions
may already have arguments and we need to take this into account.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>