when autodetecting only skip initial declarations in outer scope
When support for declarations was added in 7af3821 (add support for (single)
declarations, Tue Apr 17 09:57:35 2012 +0200), provisions were taken to
allow initial declarations to be skipped in the outer scope.
Due to a mistake in the implementation, however, _any_ declaration
would be skipped, even a declaration in the middle of the sequence
of statements.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>