separate out shared PetScan::update_scop_start_end, fixing range of DeclStmt
Essentially the same code was used in two places, one to initalize
the pet_scop corresponding to a pet_expr and one to extend the range
of the pet_scop.
The only difference between the two was that the first would
also include the semicolon after the Stmt corresponding to the pet_expr.
However, this first instance was also being used during the extraction
of a DeclStmt, the range of which already includes the semicolon,
resulting in an invalid range for those pet_scop objects.
In practice this probably didn't have any adverse effects
since the range of the pet_scop would get extended anyway.
Still, it seems more prudent to set the range correctly.

Merge the two pieces of code and make the inclusion of semicolon explicit.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>