PetScan::extract(IfStmt *): clear assignments in body
In particular, if a variable is assigned a value in the body,
then it should not be treated as a parameter in the if condition.

Note that the main purpose of the filter2 test case was to check
that the assignment of 5 to index did not end up in the condition
outside the assignment, but the test case output was wrong because
it would still treat the outer "index" as a parameter even though
it is modified inside the body.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>